ClearVC/amber-brain/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
retell-sdk==1.0.0
openai==1.6.1
langchain==0.1.0
langchain-openai==0.0.5
chromadb==0.4.18
boto3==1.34.11
twilio==8.10.1
sendgrid==6.11.0
celery==5.3.4
flower==2.0.1
prometheus-client==0.19.0
sentry-sdk==1.39.1
python-dotenv==1.0.0
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
ruff==0.1.9